reorganize tests

+++ mocpy/moc/boundaries.py
@@ -78,9 +78,9 @@ class Boundaries():
 
         # Phase 1: Retrieve the ipixels located at the border of
         # this connexe MOC component
-        ipixels = ipixels.astype(int)
+        ipixels = ipixels.astype(np.int)
         neighbours = hp.neighbours(ipixels)[[0, 2, 4, 6], :]
-        ipixels = ipixels.astype(np.uint64)
+        #ipixels = ipixels.astype(np.uint64)
         neighbours = neighbours.astype(np.uint64)
 
         isin = np.isin(neighbours, ipixels)
